+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Use new `.rebuild` API
+
 # 0.1.11
 
 * Update dependencies