+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # master
 
+# 0.1.11
+
+* Update dependencies
+
+# 0.1.10
+
+* Do not override this.inputEncoding/this.outputEncoding if not provided
+
+# 0.1.9
+
+* Fix inputEncoding/outputEncoding defaults
+
+# 0.1.8
+
+* Add `inputEncoding` and `outputEncoding` options
+
+# 0.1.7
+
+* Update dependency to deal with symlinks correctly
+
 # 0.1.6
 
 * Copy instead of hardlinking